Tooltip over scatter points show data for point
Previously, tooltip showed data for the year that the swarm was
currently representing, even if that year was not the year of the
particular scatter point being hovered over.

Now the tool tip is specific to the point being hovered over.

Added optional argument to the tt_fill function for specifying the year
of the data to be shown.  If not passed, defaults to global year.

In mouseover event listener for scatter points, year of the point is
passed in through datum passed into the call back.

Removed tooltip on hover for the lines, as doesn't quite make sense in
this new scheme.